created http-request and loader component

+++ screens/ManageExpense/ManageExpenseScreen.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useContext } from 'react';
+import React, { useEffect, useContext, useState } from 'react';
 import { StyleSheet, View } from 'react-native';
 
 import IconButton from '../../components/UI/IconButton/IconButton';
@@ -6,13 +6,16 @@ import ExpenseForm from '../../components/ManageExpense/ExpenseForm/ExpenseForm'
 
 import { GlobalStyles } from '../../constans/styles';
 import { ExpensesContext } from '../../store/expenses-context';
+import { addExpenseDb, deleteExpenseDb, updateExpenseDb } from '../../utils/http';
 
 import { IManageExpenseScreen } from './ManageExpenseScreen.props';
 import { ExpenseType } from '../../globalTypes/expenseType';
+import LoadingOverlay from '../../components/UI/LoadingOverlay/LoadingOverlay';
 
 const ManageExpenseScreen: React.FC<IManageExpenseScreen> = ({ route, navigation }) => {
   const { expenseId } = route.params;
   const isEditable = !!expenseId;
+  const [isLoading, setIsLoading] = useState<boolean>(false);
 
   const expensesCtx = useContext(ExpensesContext);
   const selectedExpense = expenseId && expensesCtx.expenses.find((exp) => exp.id === expenseId);
@@ -27,19 +30,40 @@ const ManageExpenseScreen: React.FC<IManageExpenseScreen> = ({ route, navigation
     navigation.goBack();
   };
 
-  const confirmHandle = (expense: ExpenseType): void => {
-    if (isEditable) {
-      expensesCtx.updateExpense(expense);
+  const confirmHandle = async (expense: ExpenseType): Promise<void> => {
+    setIsLoading(true);
+    if (isEditable && expense.id) {
+      const expForUpdateDb: Omit<ExpenseType, 'id'> = {
+        amount: expense.amount,
+        date: expense.date,
+        description: expense.description,
+      };
+      const status = await updateExpenseDb(expense.id, expForUpdateDb);
+      if (status === 200) {
+        expensesCtx.updateExpense(expense);
+      }
     } else {
-      expensesCtx.addExpense(expense);
+      const expenseId = await addExpenseDb(expense);
+      expensesCtx.addExpense({ ...expense, id: expenseId });
     }
+    setIsLoading(false);
     navigation.goBack();
   };
 
-  const deleteHandler = (): void => {
-    expensesCtx.deleteExpense(expenseId);
+  const deleteHandler = async (): Promise<void> => {
+    setIsLoading(true);
+    const status = await deleteExpenseDb(expenseId);
+    if (status === 200) {
+      expensesCtx.deleteExpense(expenseId);
+    }
+    setIsLoading(false);
     navigation.goBack();
   };
+
+  if (isLoading) {
+    return <LoadingOverlay />;
+  }
+
   return (
     <View style={styles.container}>
       <ExpenseForm
